+++ packages/algob/sample-project/common/algob.config.js
@@ -50,10 +50,23 @@ let defaultCfg = {
   // Below is a token created through our script in `/infrastructure`
   // If you use other setup, update it accordignly (eg content of algorand-node-data/algod.token)
   token: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+  // you can also pass token as an object
+  // token: {
+  //   "X-Algo-API-Token": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+  // },
   accounts: accounts,
   // if you want to load accounts from KMD, you need to add the kmdCfg object. Please read
   // algob-config.md documentation for details.
   // kmdCfg: kmdCfg,
+};
+
+// purestake testnet config
+let purestakeTestNetCfg = {
+  host: "https://testnet-algorand.api.purestake.io/ps2",
+  port: '',
+  token: {
+    'X-API-Key': 'Xhkn7v7h972hj7Egx3fGr9RFbfXeGuoD6wSLKDyG'
+  }
 };
 
 // You can also use Environment variables to get Algod credentials
@@ -75,6 +88,7 @@ let envCfg = {
 module.exports = {
   networks: {
     default: defaultCfg,
-    prod: envCfg
+    prod: envCfg,
+    purestake: purestakeTestNetCfg
   }
 };